refactor(api): move analytics to modules/ as controller, model, service
