+++ apps/api/src/modules/analytics/index.ts
@@ -1,7 +1,24 @@
-import { Elysia, t } from 'elysia';
-import { guards } from '../shared/guards';
-import { mcpTool } from '../mcp/generate';
-import { accessErrors, commonErrors } from '../shared/responses';
+import { Elysia } from 'elysia';
+import { guards } from '#shared/guards';
+import { mcpTool } from '#mcp/generate';
+import { accessErrors, commonErrors } from '#shared/responses';
+import {
+  ActivityPage,
+  AgentRunFeedListResponse,
+  AgentRunStatsDto,
+  AgentWorkloadListResponse,
+  BreakdownListResponse,
+  PulseListResponse,
+  StatsDto,
+  ThroughputListResponse,
+  WebhookStatsDto,
+  activityQuery,
+  agentRunFeedQuery,
+  breakdownQuery,
+  daysQuery,
+  pulseQuery,
+  throughputQuery,
+} from './model';
 import {
   getStats,
   getBreakdown,
@@ -13,119 +30,22 @@ import {
   getWebhookStats,
   getAgentWorkload,
   type ActivityCursor,
-} from './store';
-
-// Issue count stats (StatsDto from the store).
-const StatsDto = t.Object({
-  open: t.Number(),
-  inProgress: t.Number(),
-  backlog: t.Number(),
-  overdue: t.Number(),
-  unassigned: t.Number(),
-  closedLast7d: t.Number(),
-});
-
-// One breakdown bucket (BreakdownItem from the store).
-const BreakdownItem = t.Object({
-  key: t.String(),
-  label: t.String(),
-  count: t.Number(),
-  color: t.Nullable(t.String()),
-});
-
-// One pulse heatmap cell (PulseBucket from the store).
-const PulseBucket = t.Object({
-  label: t.String(),
-  count: t.Number(),
-});
-
-// One throughput week (ThroughputWeek from the store).
-const ThroughputWeek = t.Object({
-  week: t.String(),
-  created: t.Number(),
-  closed: t.Number(),
-});
-
-// One activity feed entry (ActivityItem from the store).
-const ActivityItem = t.Object({
-  id: t.Number(),
-  issueId: t.Number(),
-  issueSequence: t.Number(),
-  issueTitle: t.String(),
-  kind: t.String(),
-  actorUserId: t.Nullable(t.String()),
-  actorName: t.Nullable(t.String()),
-  body: t.Nullable(t.String()),
-  action: t.Nullable(t.String()),
-  subject: t.Nullable(t.String()),
-  fromText: t.Nullable(t.String()),
-  toText: t.Nullable(t.String()),
-  createdAt: t.String(),
-});
-
-// The keyset cursor for the next activity page (ActivityCursor from the store).
-const ActivityCursorDto = t.Object({
-  ts: t.String(),
-  id: t.Number(),
-});
-
-// One page of the activity feed (ActivityPage from the store).
-const ActivityPage = t.Object({
-  items: t.Array(ActivityItem),
-  nextCursor: t.Nullable(ActivityCursorDto),
-});
-
-// One agent run in the project-wide feed (AgentRunFeedItem from the store).
-const AgentRunFeedItem = t.Object({
-  id: t.Number(),
-  status: t.String(),
-  trigger: t.Union([
-    t.Literal('mention'),
-    t.Literal('delegation'),
-    t.Literal('schedule'),
-    t.Literal('manual'),
-  ]),
-  agentId: t.Number(),
-  agentName: t.String(),
-  issueId: t.Nullable(t.Number()),
-  issueSequence: t.Nullable(t.Number()),
-  lastError: t.Nullable(t.String()),
-  createdAt: t.String(),
-});
-
-// Agent run outcome counts over a window (AgentRunStatsDto from the store).
-const AgentRunStatsDto = t.Object({
-  total: t.Number(),
-  success: t.Number(),
-  failed: t.Number(),
-  pending: t.Number(),
-});
-
-// Webhook delivery health (WebhookStatsDto from the store).
-const WebhookStatsDto = t.Object({
-  total: t.Number(),
-  success: t.Number(),
-  failed: t.Number(),
-  pending: t.Number(),
-  activeWebhooks: t.Number(),
-  disabledWebhooks: t.Number(),
-});
-
-// One agent's workload row (AgentWorkloadItem from the store).
-const AgentWorkloadItem = t.Object({
-  agentId: t.Number(),
-  agentName: t.String(),
-  kind: t.String(),
-  delegatedOpen: t.Number(),
-  runsTotal: t.Number(),
-  runsSuccess: t.Number(),
-  runsFailed: t.Number(),
-});
+  type PulseUnit,
+} from './service';
 
 // Read-only project analytics that back the dashboard widgets. Every route is
 // under /projects/:projectKey/analytics and gated by the dashboards read
 // permission (the analytics feed the dashboards). All figures come from existing
-// tables; see store.ts.
+// tables; see service.ts.
+
+// Caps the generated bucket axis per unit. Hour columns hold 24 cells each, so
+// their cap is the lowest.
+const MAX_PULSE_COLUMNS: Record<PulseUnit, number> = {
+  hour: 140,
+  day: 160,
+  week: 130,
+};
+
 export const analyticsRoutes = new Elysia({
   name: 'analytics',
   detail: { tags: ['Analytics'] },
@@ -153,17 +73,9 @@ export const analyticsRoutes = new Elysia({
       return getBreakdown(project.id, query.by);
     },
     {
-      query: t.Object({
-        by: t.Union([
-          t.Literal('status'),
-          t.Literal('priority'),
-          t.Literal('type'),
-          t.Literal('assignee'),
-          t.Literal('delegate'),
-        ]),
-      }),
+      query: breakdownQuery,
       permission: ['dashboards', 'read'],
-      response: { 200: t.Array(BreakdownItem), ...commonErrors },
+      response: { 200: BreakdownListResponse, ...commonErrors },
       detail: {
         summary: 'Get project breakdown',
         description: 'Issue counts grouped by a chosen dimension.',
@@ -176,19 +88,14 @@ export const analyticsRoutes = new Elysia({
     '/projects/:projectKey/analytics/pulse',
     async ({ project, query }) => {
       const unit = query.unit ?? 'day';
-      // Cap columns so the generated bucket axis stays bounded (hour columns hold
-      // 24 cells each, so their cap is lower).
-      const maxColumns = unit === 'hour' ? 140 : unit === 'week' ? 130 : 160;
-      const columns = query.columns != null ? Math.min(Math.max(query.columns, 1), maxColumns) : 26;
+      const columns =
+        query.columns != null ? Math.min(Math.max(query.columns, 1), MAX_PULSE_COLUMNS[unit]) : 26;
       return getPulse(project.id, unit, columns);
     },
     {
-      query: t.Object({
-        unit: t.Optional(t.Union([t.Literal('hour'), t.Literal('day'), t.Literal('week')])),
-        columns: t.Optional(t.Numeric()),
-      }),
+      query: pulseQuery,
       permission: ['dashboards', 'read'],
-      response: { 200: t.Array(PulseBucket), ...commonErrors },
+      response: { 200: PulseListResponse, ...commonErrors },
       detail: {
         summary: 'Get project pulse',
         description: 'Activity counts over time for a heatmap.',
@@ -204,9 +111,9 @@ export const analyticsRoutes = new Elysia({
       return getThroughput(project.id, weeks);
     },
     {
-      query: t.Object({ weeks: t.Optional(t.Numeric()) }),
+      query: throughputQuery,
       permission: ['dashboards', 'read'],
-      response: { 200: t.Array(ThroughputWeek), ...commonErrors },
+      response: { 200: ThroughputListResponse, ...commonErrors },
       detail: {
         summary: 'Get project throughput',
         description: 'Created versus closed issues over time.',
@@ -243,15 +150,7 @@ export const analyticsRoutes = new Elysia({
       });
     },
     {
-      query: t.Object({
-        limit: t.Optional(t.Numeric()),
-        cursor: t.Optional(t.String()),
-        actorUserId: t.Optional(t.String()),
-        action: t.Optional(t.String()),
-        // Comma-separated issue ids the client resolved from the widget's work items
-        // filter; absent means no issue-scope restriction.
-        issueIds: t.Optional(t.String()),
-      }),
+      query: activityQuery,
       permission: ['dashboards', 'read'],
       response: { 200: ActivityPage, ...commonErrors },
       detail: {
@@ -269,14 +168,9 @@ export const analyticsRoutes = new Elysia({
       return listAgentRunFeed(project.id, { status: query.status ?? null, limit });
     },
     {
-      query: t.Object({
-        status: t.Optional(
-          t.Union([t.Literal('pending'), t.Literal('success'), t.Literal('failed')]),
-        ),
-        limit: t.Optional(t.Numeric()),
-      }),
+      query: agentRunFeedQuery,
       permission: ['dashboards', 'read'],
-      response: { 200: t.Array(AgentRunFeedItem), ...commonErrors },
+      response: { 200: AgentRunFeedListResponse, ...commonErrors },
       detail: {
         summary: 'List agent runs',
         description: 'Project-wide feed of agent runs.',
@@ -292,7 +186,7 @@ export const analyticsRoutes = new Elysia({
       return getAgentRunStats(project.id, days);
     },
     {
-      query: t.Object({ days: t.Optional(t.Numeric()) }),
+      query: daysQuery,
       permission: ['dashboards', 'read'],
       response: { 200: AgentRunStatsDto, ...commonErrors },
       detail: {
@@ -310,7 +204,7 @@ export const analyticsRoutes = new Elysia({
       return getWebhookStats(project.id, days);
     },
     {
-      query: t.Object({ days: t.Optional(t.Numeric()) }),
+      query: daysQuery,
       permission: ['dashboards', 'read'],
       response: { 200: WebhookStatsDto, ...commonErrors },
       detail: {
@@ -328,7 +222,7 @@ export const analyticsRoutes = new Elysia({
     },
     {
       permission: ['dashboards', 'read'],
-      response: { 200: t.Array(AgentWorkloadItem), ...accessErrors },
+      response: { 200: AgentWorkloadListResponse, ...accessErrors },
       detail: {
         summary: 'Get agent workload',
         description: 'Per-agent open delegated issues and run outcomes.',
