+++ apps/api/src/modules/analytics/service.ts
@@ -11,7 +11,7 @@ import {
   webhookDelivery,
 } from '@repo/db';
 import { and, desc, eq, inArray, isNull, isNotNull, sql } from 'drizzle-orm';
-import { iso } from '../shared/lib';
+import { iso } from '#shared/lib';
 
 // Read-only project metrics for the dashboards feature. Every figure is derived
 // from the existing issue / project_column / issue_activity tables — there is no
